Simple (but probably old way) spring cloud stream rabbitmq communication

+++ organization-service/src/main/java/com/learning/eagleeye/organization/service/OrganizationService.java
@@ -1,5 +1,6 @@
 package com.learning.eagleeye.organization.service;
 
+import com.learning.eagleeye.organization.events.source.SimpleSourceBean;
 import com.learning.eagleeye.organization.model.Organization;
 import com.learning.eagleeye.organization.repository.OrganizationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,10 +14,12 @@ import java.util.UUID;
 public class OrganizationService {
 
     private OrganizationRepository organizationRepository;
+    private SimpleSourceBean simpleSourceBean;
 
     @Autowired
-    public OrganizationService(OrganizationRepository organizationRepository) {
+    public OrganizationService(OrganizationRepository organizationRepository, SimpleSourceBean simpleSourceBean) {
         this.organizationRepository = organizationRepository;
+        this.simpleSourceBean = simpleSourceBean;
     }
 
     public Optional<Organization> get(String organizationId) {
@@ -30,15 +33,20 @@ public class OrganizationService {
     public void save(Organization organization) {
         organization.setId(UUID.randomUUID().toString());
         organizationRepository.save(organization);
+        simpleSourceBean.publishOrganizationChange("SAVE", organization.getId());
     }
 
     public void update(Organization organization) {
         organizationRepository.save(organization);
+        simpleSourceBean.publishOrganizationChange("UPDATE", organization.getId());
     }
 
     public void delete(String organizationId) {
         try {
-            organizationRepository.deleteById(organizationId);
+            if (organizationRepository.findById(organizationId).isPresent()) {
+                organizationRepository.deleteById(organizationId);
+                simpleSourceBean.publishOrganizationChange("DELETE", organizationId);
+            }
         } catch (EmptyResultDataAccessException ignored) { }
     }
 }
